refactor: change rerank_factor from uint32/usize to float/f32
Allow fractional rerank factors (e.g. 1.5x, 2.5x candidates).
All fetch_k computations now use ceil(k * rerank_factor) for rounding.

Changes:
- proto: DenseVectorQuery.rerank_factor uint32 -> float
- core: DenseVectorQuery.rerank_factor usize -> f32, default 3.0
- core: segment reader fetch_k = ceil(k * rerank_factor.max(1.0))
- core: QL parser rerank param parsed as f32
- server: converter passes f32 directly (no usize cast)
- python: DenseVectorQuery.rerank_factor type int -> float
- typescript: regenerated proto stubs (rerankFactor already number)

+++ hermes-client-python/src/hermes_client_python/types.py
@@ -62,7 +62,7 @@ class DenseVectorQuery(TypedDict, total=False):
     field: str  # required but total=False for optional fields
     vector: list[float]
     nprobe: int
-    rerank_factor: int
+    rerank_factor: float
     combiner: Combiner
     combiner_temperature: float
     combiner_top_k: int
